Allow to reuse the dockerfilegen from other packages
