+++ pkg/dockerfilegen/dockerfile-templates/BuildImage.dockerfile.tmpl
