+++ pkg/dockerfilegen/dockerfile-templates/Default.dockerfile.tmpl
