+++ pkg/dockerfilegen/dockerfile-templates/FuncUtil.dockerfile.tmpl
