+++ pkg/dockerfilegen/dockerfile-templates/MustGather.dockerfile.tmpl
