+++ pkg/dockerfilegen/dockerfile-templates/SourceImage.dockerfile.tmpl
